Extract storage operation functions from cloud_utils.py into storage_operations.py
